Move logo out into DesktopGrid
