+++ src/components/Layout/DesktopGrid.module.css
@@ -87,23 +87,26 @@
   align-items: center;
 }
 
-@keyframes background {
-  0% {
-    background-color: #ffffff;
+.logo {
+  font-size: 3.75rem;
+  margin: 0 auto 0;
+  opacity: 1;
+  position: sticky;
+  text-align: center;
+  top: 3.5rem;
+  width: 100%;
+  z-index: 1;
+
+  &.hidden {
+    opacity: 0;
   }
-  20% {
-    background-color: #e3ffb5;
-  }
-  40% {
-    background-color: #e57200;
-  }
-  60% {
-    background-color: #5dade2;
-  }
-  80% {
-    background-color: #cc8899;
-  }
-  100% {
-    background-color: #ffffff;
-  }
+}
+
+.hugeLogo {
+  display: block;
+  font-size: 10vw;
+  height: 100vh;
+  line-height: 100vh;
+  margin: 0 auto;
+  text-align: center;
 }
