+++ src/components/Layout/Layout.module.css
@@ -1,27 +1,3 @@
-.logo {
-  font-size: 3.75rem;
-  margin: 0 auto 0;
-  opacity: 1;
-  position: sticky;
-  text-align: center;
-  top: 3.5rem;
-  width: 100%;
-  z-index: 1;
-
-  &.hidden {
-    opacity: 0;
-  }
-}
-
-.hugeLogo {
-  display: block;
-  font-size: 10vw;
-  height: 100vh;
-  line-height: 100vh;
-  margin: 0 auto;
-  text-align: center;
-}
-
 .footer {
   font-size: 1.5rem;
   height: 12.5rem;
