+++ src/components/Layout/Layout.tsx
@@ -5,59 +5,25 @@ import styles from "./Layout.module.css";
 
 import { Language } from "../../types";
 import { DesktopGrid } from "./DesktopGrid";
+import { useWindowSize } from "react-use";
+import { MobileWrapper } from "./MobileWrapper";
 
 export function Layout() {
   const { lang } = useParams() as { lang: Language };
+  const { width } = useWindowSize();
+  const isMobile = width < 1024;
 
   const date = new Date();
-  const [intersects, setIntersects] = useState(true);
 
-  const location = useLocation();
-  // eslint-disable-next-line
-  const isHome = location.pathname.match(/^\/(pl|en)[\/]?$/);
-  const intersectionRef = useRef<HTMLHeadingElement>(null);
-  useEffect(() => {
-    let observer: any;
-    if (isHome) {
-      observer = new IntersectionObserver(
-        (entries) => {
-          setIntersects(entries.slice(-1)[0].isIntersecting);
-        },
-        {
-          root: null,
-          rootMargin: "0px",
-          threshold: 0.5,
-        },
-      );
-      if (intersectionRef.current) {
-        observer.observe(intersectionRef.current);
-      }
-    } else {
-      setIntersects(false);
-    }
-
-    return () => observer?.disconnect();
-  }, [isHome, intersects]);
+  const Wrapper = isMobile ? MobileWrapper : DesktopGrid;
 
   return (
-    <DesktopGrid lang={lang}>
-      <h1
-        className={classNames(styles.logo, {
-          [styles.hidden]: intersects,
-        })}
-      >
-        problonde
-      </h1>
-      {isHome && (
-        <h1 className={styles.hugeLogo} ref={intersectionRef}>
-          problonde
-        </h1>
-      )}
+    <Wrapper lang={lang}>
       <Outlet />
       <div className={styles.footer}>
         created by Problonde Studio
         {date.getFullYear()}
       </div>
-    </DesktopGrid>
+    </Wrapper>
   );
 }
