+++ src/components/Layout/MobileWrapper.module.css
@@ -1,0 +1,12 @@
+.wrapper {
+  display: block;
+  font-family: "General Sans", sans-serif;
+  background-color: var(--layout-bg-color);
+  animation-name: none;
+  animation-duration: 10s;
+  animation-timing-function: linear;
+  animation-iteration-count: infinite;
+  animation-play-state: running;
+
+  transition: all 0.2s;
+}
